server: Added netascii mode
netascii is a silly mode from the 70's when you would use tftp over
telephone lines. So there is a reader now that a transpose your input
stream from octet (the mode with bytes of 8 bit each) to netascii, with
is byte allinged as well but doesn't use one of those bytes.

The docuementation states that this is important for 36 bit DEC machines
but some clients default to this mode. (some server might still be 36
bit machines from DEC)